Rename LimitHeadingsExtension to NormalizeHeadingsExtension
The "Limit" name bakes clamping into the API, leaving no room for the
heading-level renumbering requested in #1115. A neutral name lets both
features live in one extension with a defined ordering.
