+++ docs/2.x/extensions/overview.md
@@ -40,8 +40,8 @@ to enhance your experience out-of-the-box depending on your specific use-cases.
 | [Heading Permalinks] | Makes heading elements linkable | `1.4.0` | |
 | [Highlight] | Mark text as being highlighted for reference or notation purposes | `2.8.0` | |
 | [Inlines Only] | Only includes standard CommonMark inline elements - perfect for handling comments and other short bits of text where you only want bold, italic, links, etc. | `1.3.0` | |
-| [Limit Headings] | Constrains heading levels to a configured min/max range | `2.9.0` | |
 | [Mentions] | Easy parsing of `@mention` and `#123`-style references | `1.5.0` | |
+| [Normalize Headings] | Constrains heading levels to a configured min/max range | `2.9.0` | |
 | [Strikethrough] | Allows using tilde characters (`~~`) for ~strikethrough~ formatting | `1.3.0`  | <i class="fab fa-github"></i> |
 | [Tables] | Enables you to create HTML tables | `1.3.0`  | <i class="fab fa-github"></i> |
 | [Table of Contents] | Automatically inserts links to the headings at the top of your document | `1.4.0` | |
@@ -123,8 +123,8 @@ See the [Custom Extensions](/2.x/customization/extensions/) page for details on 
 [Heading Permalinks]: /2.x/extensions/heading-permalinks/
 [Highlight]: /2.x/extensions/highlight/
 [Inlines Only]: /2.x/extensions/inlines-only/
-[Limit Headings]: /2.x/extensions/limit-headings/
 [Mentions]: /2.x/extensions/mentions/
+[Normalize Headings]: /2.x/extensions/normalize-headings/
 [Strikethrough]: /2.x/extensions/strikethrough/
 [Tables]: /2.x/extensions/tables/
 [Table of Contents]: /2.x/extensions/table-of-contents/
