+++ docs/_data/menu.yml
@@ -27,8 +27,8 @@ version:
             'Heading Permalinks': '/2.x/extensions/heading-permalinks/'
             'Highlight': '/2.x/extensions/highlight/'
             'Inlines Only': '/2.x/extensions/inlines-only/'
-            'Limit Headings': '/2.x/extensions/limit-headings/'
             'Mentions': '/2.x/extensions/mentions/'
+            'Normalize Headings': '/2.x/extensions/normalize-headings/'
             'Smart Punctuation': '/2.x/extensions/smart-punctuation/'
             'Strikethrough': '/2.x/extensions/strikethrough/'
             'Table of Contents': '/2.x/extensions/table-of-contents/'
