+++ tests/functional/Extension/NormalizeHeadings/md/clamps-levels.md
@@ -1,5 +1,5 @@
 ---
-limit_headings:
+normalize_headings:
     min_heading_level: 2
     max_heading_level: 4
 ---
